Add structure of program
